# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/appointment-services/staff/{staffId}": {
      "get": {
        "operationId": "AppointmentServicesController_findAppointmentByStaff_v1",
        "parameters": [
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentService"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Appointmentservices"]
      }
    },
...
    "/api/v1/schedules/staff/{staffId}": {
      "get": {
        "operationId": "SchedulesController_findByStaff_v1",
        "parameters": [
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Schedules"]
      }
    },
    "/api/v1/staffs": {
      "post": {
        "operationId": "StaffsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Staffs"]
      },
      "get": {
        "operationId": "StaffsController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfinityPaginationStaffResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Staffs"]
      }
    },
    "/api/v1/staffs/user/{id}": {
      "get": {
        "operationId": "StaffsController_findByUser_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Staffs"]
      }
    },
    "/api/v1/staffs/{id}": {
      "get": {
        "operationId": "StaffsController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Staffs"]
      },
      "patch": {
        "operationId": "StaffsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Staffs"]
      },
      "delete": {
        "operationId": "StaffsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Staffs"]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FileType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae"
          },
          "path": {
            "type": "string",
            "example": "https://example.com/path/to/file.jpg"
          }
        },
        "required": ["id", "path"]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string",
            "example": "admin"
          }
        },
        "required": ["id", "name"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "zaloId": {
            "type": "string",
            "example": "1234567890"
          },
          "phoneNumber": {
            "type": "string",
            "example": "1234567890"
          },
          "userName": {
            "type": "string",
            "example": "John"
          },
          "avatar": {
            "example": "avatar image link",
            "allOf": [
              {
                "$ref": "#/components/schemas/FileType"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "zaloId",
          "phoneNumber",
          "userName",
          "avatar",
          "role",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "StatusEnum": {
        "type": "number",
        "enum": [1, 2, 3, 4, 5, 6, 7, 8],
        "description": "Status of the appointment (1 = pending, 2 = inprogress, 3 = done)"
      },
      "CustomerRecord": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "nullable": false
          },
          "BHYTNumber": {
            "type": "string",
            "nullable": true
          },
          "CCCDNumber": {
            "type": "string",
            "nullable": true
          },
          "DOB": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "sex": {
            "type": "string",
            "nullable": false
          },
          "fullName": {
            "type": "string",
            "nullable": false
          },
          "user": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "active",
          "BHYTNumber",
          "CCCDNumber",
          "DOB",
          "sex",
          "fullName",
          "user",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "specificTime": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "active": {
            "type": "boolean",
            "nullable": false
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "description": "Status of the appointment (1 = pending, 2 = inprogress, 3 = done)",
            "example": 1,
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusEnum"
              }
            ]
          },
          "customerRecord": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/CustomerRecord"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "specificTime",
          "active",
          "note",
          "status",
          "customerRecord",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "AppointmentService": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "string",
            "nullable": false
          },
          "appointment": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/Appointment"
              }
            ]
          },
          "serviceId": {
            "type": "string",
            "nullable": false
          },
          "staffId": {
            "type": "string",
            "nullable": false
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "scheduleId",
          "appointment",
          "serviceId",
          "staffId",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Staff": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          },
          "specialization": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "user": {
            "example": "userId",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "note",
          "specialization",
          "id",
          "user",
          "createdAt",
          "updatedAt"
        ]
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean",
            "nullable": false
          },
          "endTime": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "startTime": {
            "format": "date-time",
            "type": "string",
            "nullable": false
          },
          "staff": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/Staff"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "note",
          "active",
          "endTime",
          "startTime",
          "staff",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateStaffDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          }
        }
      },
      "InfinityPaginationStaffResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Staff"
            }
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["data", "hasNextPage"]
      },
      "UpdateStaffDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          }
        }
      },
...
    }
  }
}